Move plugins into packages per domain, share interfaces between plugins
