+++ plugins/stores.go
@@ -1,0 +1,10 @@
+package plugins
+
+import types "code.vegaprotocol.io/vega/proto"
+
+// CandleStore persistence for candles
+//go:generate go run github.com/golang/mock/mockgen -destination mocks/candle_store_mock.go -package mocks code.vegaprotocol.io/vega/plugins CandleStore
+type CandleStore interface {
+	FetchLastCandle(marketID string, interval types.Interval) (*types.Candle, error)
+	GenerateCandlesFromBuffer(marketID string, previousCandlesBuf map[string]types.Candle) error
+}
